Start adding support for an errorprone checker.
By using errorprone, we can provide suggested fixes for many of the
issues detected by the spotbugsPlugin (and do it in the normal
compilation as well). Since errorprone also allows one to inline patch,
we should be able to automatically patch many cases.

The tests should be broken out _per_ avro version as the compiler is
actually run with the given classpath. We _could_ try to somehow
override the classpath, by say, adding all the jar files needed for
testing to the compiler helpers, but that may be more intrusive than
using different gradle projects.